+++ src/domain/errors/TrailerCodecError.js
@@ -4,13 +4,11 @@
 export default class TrailerCodecError extends Error {
   /**
    * @param {string} message - Human readable error message.
-   * @param {string} code - Machine readable error code.
    * @param {Object} [meta] - Additional metadata context.
    */
-  constructor(message, code, meta = {}) {
+  constructor(message, meta = {}) {
     super(message);
     this.name = this.constructor.name;
-    this.code = code;
     this.meta = meta;
     Error.captureStackTrace(this, this.constructor);
   }